requirements.txt
numpy
pandas
PyYAML
stomp.py
mysql-connector-python
pytz
pymongo
nose